+++ lib/language_pack/helpers/download_presence.rb
@@ -23,6 +23,7 @@ class LanguagePack::Helpers::DownloadPresence
     @file_name = file_name
 
     # Special case: Ruby 2.6.6 is only available on heroku-20, not heroku-22
+    # Check heroku-20 availability when requesting heroku-22
     if file_name == "ruby-2.6.6.tgz"
       @stacks = stacks.map { |stack| stack == "heroku-22" ? "heroku-20" : stack }
     else